Move React pages to pages folder
