+++ client/routes.jsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
-import Login from '../imports/ui/components/login.jsx';
-import Landing from '../imports/ui/components/landing.jsx';
+import Login from '../imports/ui/pages/login.jsx/index.js';
+import Landing from '../imports/ui/pages/landing.jsx/index.js';
 import Navbar from '../imports/ui/components/navbar.jsx';
-import Faq from '../imports/ui/components/faq.jsx';
-import Woohoo from '../imports/ui/components/woohoo.jsx';
-import Apply from '../imports/ui/components/apply.jsx';
-import ForgotPassword from '../imports/ui/components/forgot_password.jsx';
-import SetPassword from '../imports/ui/components/set_password.jsx';
-import Moderator from '../imports/ui/components/moderator.jsx';
+import Faq from '../imports/ui/pages/faq.jsx';
+import Woohoo from '../imports/ui/pages/woohoo.jsx/index.js';
+import Apply from '../imports/ui/pages/apply.jsx/index.js';
+import ForgotPassword from '../imports/ui/pages/forgot_password.jsx/index.js';
+import SetPassword from '../imports/ui/pages/set_password.jsx/index.js';
+import Moderator from '../imports/ui/pages/moderator.jsx/index.js';
 import Profile from '../imports/ui/components/dashboard/profile';
 import AuthenticatedRoute from '../imports/ui/hoc/AuthenticatedRoute';
 import { AuthProvider } from '../imports/ui/hoc/AuthProvider';
